Fixing Auth0LockPasswordless export
